moved conv-family commands into its own system plugin
